requirements.txt
Django>=5.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
mysql-connector-python>=9.0.0
scikit-learn>=1.3.0
openai>=1.0.0
google-generativeai>=0.3.0
